# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jewellery Shop API",
        "description": "Jewellery Shop Management API",
        "termsOfService": "https://jewellery-shop.test/terms",
        "contact": {
            "email": "admin@jewellery-shop.test"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/backups": {
            "get": {
                "tags": [
                    "Backups"
                ],
                "summary": "List backups",
                "description": "Return a list of all available database backups",
                "operationId": "88046cf90195dfae8e66d6210d25a7d2",
                "responses": {
                    "200": {
                        "description": "List of backups",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "keep_days": {
                                                    "type": "integer",
                                                    "example": 7
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Backups"
                ],
                "summary": "Create backup",
                "description": "Run a new database backup",
                "operationId": "7219e360b3d165655516c5e3811727dd",
                "responses": {
                    "200": {
                        "description": "Backup created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "backup-2024-01-01.sql.gz"
                                                },
                                                "size": {
                                                    "type": "integer",
                                                    "example": 102400
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "tables": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "rows": {
                                                    "type": "integer",
                                                    "example": 5000
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Backup completed."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/backups/{name}/download": {
            "get": {
                "tags": [
                    "Backups"
                ],
                "summary": "Download backup",
                "description": "Download a specific backup file",
                "operationId": "758818a5d20f6f268d47023910dd1e24",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Backup filename",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Backup file download",
                        "content": {
                            "application/gzip": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Backup not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/backups/{name}": {
            "delete": {
                "tags": [
                    "Backups"
                ],
                "summary": "Delete backup",
                "description": "Delete a specific backup file",
                "operationId": "c9a98f886583590b83de3d059c4db6a1",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Backup filename",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Backup deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Backup not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}